Get dynamic routing working
This part was a beast and required changing from a layout
situation to putting it in the slug and dynamically rendering it. Not
honestly super sure how this is working, but it is working. Hooray I
guess. Was it just simpler before? Yeah but now I can so easily make and
iterate on new blog entries.